Rearrange of project
