+++ app/v1/global_product/__init__.py
@@ -1,4 +1,7 @@
-from .model import *
+from .model1 import *
+from .controller1 import *
+from .route1 import *
+# from app.User.model import *
 from .controller import *
 from .route import *
-# from app.User.model import *
+from .model import *
