+++ app/v1/global_product/controller/__init__.py
@@ -1,0 +1,3 @@
+from .category import *
+from .features import *
+from .product import *
