+++ app/v1/global_product/model/category/__init__.py
@@ -1,0 +1,1 @@
+from .category import *
