+++ app/v1/global_product/model/features/__init__.py
@@ -1,0 +1,1 @@
+from .features import *
