+++ app/v1/global_product/model/product/__init__.py
@@ -1,0 +1,1 @@
+from .product import *
