+++ app/v1/global_product/route/category/__init__.py
@@ -1,0 +1,4 @@
+from .add import *
+from .fetch import *
+from .edit import *
+from .delete import *
